pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robo-trader-agent"
version = "0.1.0"
description = "LLM-powered automated trading system for crypto and equities via Alpaca"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.21.0",
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "ta>=0.11.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "loguru>=0.7.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
